Cleaned up; moved parseOperation to separate file
